requirements.txt
numpy==1.25.2
pandas==2.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
tzdata==2023.3
